+++ xray/core/variable.py
@@ -15,6 +15,11 @@ from .indexing import (PandasIndexAdapter, orthogonally_indexable)
 
 import xray  # only for Dataset and DataArray
 
+try:
+    import dask.array as da
+except ImportError:
+    pass
+
 
 def as_variable(obj, key=None, strict=True):
     """Convert an object into an Variable
@@ -494,6 +499,110 @@ class Variable(common.AbstractArray, utils.NdimSizeLenMixin):
             if dim in indexers:
                 key[i] = indexers[dim]
         return self[tuple(key)]
+
+    def _shift_one_dim(self, dim, count):
+        axis = self.get_axis_num(dim)
+
+        if count > 0:
+            keep = slice(None, -count)
+        elif count < 0:
+            keep = slice(-count, None)
+        else:
+            keep = slice(None)
+
+        trimmed_data = self[(slice(None),) * axis + (keep,)].data
+        dtype, fill_value = common._maybe_promote(self.dtype)
+
+        shape = list(self.shape)
+        shape[axis] = min(abs(count), shape[axis])
+
+        if isinstance(trimmed_data, dask_array_type):
+            chunks = list(trimmed_data.chunks)
+            chunks[axis] = (shape[axis],)
+            full = functools.partial(da.full, chunks=chunks)
+        else:
+            full = np.full
+
+        nans = full(shape, fill_value, dtype=dtype)
+
+        if count > 0:
+            arrays = [nans, trimmed_data]
+        else:
+            arrays = [trimmed_data, nans]
+
+        data = ops.concatenate(arrays, axis)
+
+        if isinstance(data, dask_array_type):
+            # chunked data should come out with the same chunks; this makes
+            # it feasible to combine shifted and unshifted data
+            # TODO: remove this once dask.array automatically aligns chunks
+            data = data.rechunk(self.data.chunks)
+
+        return type(self)(self.dims, data, self._attrs, fastpath=True)
+
+    def shift(self, **shifts):
+        """
+        Return a new Variable with shifted data.
+
+        Parameters
+        ----------
+        **shifts : keyword arguments of the form {dim: offset}
+            Integer offset to shift along each of the given dimensions.
+            Positive offsets shift to the right; negative offsets shift to the
+            left.
+
+        Returns
+        -------
+        shifted : Variable
+            Variable with the same dimensions and attributes but shifted data.
+        """
+        result = self
+        for dim, count in shifts.items():
+            result = result._shift_one_dim(dim, count)
+        return result
+
+    def _roll_one_dim(self, dim, count):
+        axis = self.get_axis_num(dim)
+
+        count %= self.shape[axis]
+        if count != 0:
+            indices = [slice(-count, None), slice(None, -count)]
+        else:
+            indices = [slice(None)]
+
+        arrays = [self[(slice(None),) * axis + (idx,)].data
+                  for idx in indices]
+
+        data = ops.concatenate(arrays, axis)
+
+        if isinstance(data, dask_array_type):
+            # chunked data should come out with the same chunks; this makes
+            # it feasible to combine shifted and unshifted data
+            # TODO: remove this once dask.array automatically aligns chunks
+            data = data.rechunk(self.data.chunks)
+
+        return type(self)(self.dims, data, self._attrs, fastpath=True)
+
+    def roll(self, **shifts):
+        """
+        Return a new Variable with rolld data.
+
+        Parameters
+        ----------
+        **shifts : keyword arguments of the form {dim: offset}
+            Integer offset to roll along each of the given dimensions.
+            Positive offsets roll to the right; negative offsets roll to the
+            left.
+
+        Returns
+        -------
+        shifted : Variable
+            Variable with the same dimensions and attributes but rolled data.
+        """
+        result = self
+        for dim, count in shifts.items():
+            result = result._roll_one_dim(dim, count)
+        return result
 
     def transpose(self, *dims):
         """Return a new Variable object with transposed dimensions.